backend/requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
